[MOV] tools: move runtime analysis in `safe_eval.runtime.py`
The `safe_eval.evaluation.py` module contains the static and dynamic
analysis features.
Since the file is becoming quite large, it makes sense to split it by
moving the dynamic analysis feature into a new module named
`safe_eval.runtime.py`.

Task-6120890
