+++ odoo/tools/safe_eval/__init__.py
@@ -1,2 +1,3 @@
-from .evaluation import *
+from .evaluation import *  # noqa: F403
 from .expression import expr_eval
+from .runtime import *  # noqa: F403
